Module lg_webos_client::client

Client Module

Current, An WebOs Client must implement two traits, SendLgCommandRequest , SendPointerCommandRequest, with SendPointerCommandRequest, the Client can control the WebOs’s Mouse, The SendLgCommandRequest is the send LGCommandRequest capability. You can do a lot with it. See the all LGCommandRequest implementations in lg_command. The WebOsClient implements the two traits.

Structs

WebOsClient

Client for interacting with TV. This client only can used if the connection is successul,

WebOsClientConfig

Parameters To connect to TV if the Client isn’t authenticated then the key must be None The address must contain the port,and the connection type. The Default port is 3000 The default connection is Websocket: ws:// If you Operational System has mDns the ip can be: lgwebostv.local

Enums

WebSocketError

Error on WebSocket Connection There are two kinds of erros: Fatal and Continue

Traits

SendLgCommandRequest

The Trait Responsable to Send LGCommandRequest to TV.
See All LGCommandRequest in lg_command module

SendPointerCommandRequest

The Trait Responsable to Send PointerInputCommand to TV.
See All PointerInputCommand in lg_command module

Attribute Macros

async_trait