Module lg_webos_client::lg_command

LG Command Module

This module has all commands that can be sended to the WebOs TV. The CommandRequest is the expected json that’s what the TV waiting for and the CommandResponse is the Json that the client expected. The Command response and Command request is sended though the main WebSocket connection. There Are a lot of commands that the TV can receive, so I created the abstraction:

the LGCommandRequest trait This trait just return a CommandRequest. See all commands is see all the structs that implements the trait. These structs is in request_commands module.

Modules

pointer_input_commands

Pointer Inputs Commands Module

request_commands

Request Commnads Module

Structs

CommandRequest

The Json that WebOs TV expects to receive

CommandResponse

The Response Expeceted form Request, but currently isn’t used

Constants

REQUEST_TYPE

Traits

LGCommandRequest

Trait that A WebOs Client Must to implement to send commands to TV