Module lg_webos_client::wake_on_lan

Wake On Lan module: A Protocol to turn on the TV

To Turn on the TV you can use the Wake On Lan Protocol To turn on the TV you will need to know the MAC address, and the IP

Turn On Example:

use lg_webos_client::wake_on_lan;

async fn turn_on_tv(ip: &str, mac: &str) {
    let magic_package = wake_on_lan::MagicPacket::from_mac_string(mac).unwrap();
    wake_on_lan::send_magic_packet_to_address(magic_package, &format!("{ip}:9")).await
        .unwrap();
}

turn_on_tv("192.168.0.199","a4:a2:22:a7:ff:3f");

Structs

MagicPacket

Functions

send_magic_packet_to_address

Sends the magic packet to especific address, expected that address has port 9

send_magic_packet_to_broadcast

Sends the magic packet to default broadcast address Maybe the TV will not receive the packet

string_to_bytes_mac_address